FEATURE: 마이프로필 조회 컨트롤러에 API 메서드 추가
